# c61dc456-3c60-47ad-b5eb-f110ba062621/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "文体场馆在线预约系统 API",
        "description": "某市公共文化服务中心在线预约系统RESTful API文档，支持实名预约、分时段计费、信用积分管理、黑名单自动拉黑",
        "contact": {
            "name": "API Support",
            "email": "admin@booking.example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/payments": {
            "post": {
                "tags": [
                    "支付管理"
                ],
                "summary": "创建支付订单",
                "operationId": "a6408a9b37441f15d1c4c5e79029ac39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "description": "预约ID",
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "description": "支付方式",
                                        "type": "string",
                                        "enum": [
                                            "wechat",
                                            "alipay"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误或业务失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未认证"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{paymentNo}/callback": {
            "post": {
                "tags": [
                    "支付管理"
                ],
                "summary": "支付回调",
                "operationId": "8ee34cca045e5420f68ba00e5b371f4c",
                "parameters": [
                    {
                        "name": "paymentNo",
                        "in": "path",
                        "description": "支付编号",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "回调处理失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/{paymentNo}/mock-pay": {
            "get": {
                "tags": [
                    "支付管理"
                ],
                "summary": "模拟支付（仅测试）",
                "operationId": "ee18ccc43fdec9be3de18745bb5ea9bd",
                "parameters": [
                    {
                        "name": "paymentNo",
                        "in": "path",
                        "description": "支付编号",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "模拟支付失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/{paymentId}/refund": {
            "post": {
                "tags": [
                    "支付管理"
                ],
                "summary": "申请退款",
                "operationId": "e36e95261d297bc42baf066e35316ba5",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "支付ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "reason"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "退款金额",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01
                                    },
                                    "reason": {
                                        "description": "退款原因",
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "退款失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未认证"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/{paymentId}/status": {
            "get": {
                "tags": [
                    "支付管理"
                ],
                "summary": "查询支付状态",
                "operationId": "67923b3123a197fa68af89709e3a69af",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "支付ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "payment_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "payment_no": {
                                                    "type": "string",
                                                    "example": "PAY20240101000001"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "paid"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                },
                                                "refund_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0,
                                                    "nullable": true
                                                },
                                                "refunded_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未认证"
                    },
                    "404": {
                        "description": "支付记录不存在"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token 认证",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}